Introduce editing primitives for pom.xml files
Demonstrate deleting blocks out of pom.xml files and prepending to the
end of pom.xml files.

XML processing alternatives included Ruby nokogiri, `yq', and rolling my
own tiny Java program.

Nokogiri doesn't do well with `.xpath' for pom.xml files with my attempt
Not saying it's impossible, but ran out of time box for it.

yq supports XML munging but cannot handle whitespace. We need the
pom.xml file to be human-readable for further processing.

And reading the JDK docs for XML processing would involve writing a lot
of classes, bootstrapping a project, build tool (Maven or Gradle), etc.
Even with Groovy, it's still another piece of tech.

We're treating the pom.xml file like a Gherkin feature file
(line-oriented), and our processing is naive. So we don't need the big
guns yet.

A sample pom.xml is included. It's not meant to be a production version,
but just to demonstrate the nex transforms.

Other observations in the demo:

* A new pattern of `read', edit, `w!'
* When the above is used, `1,$d' is followed by `w /tmp/blah' and
  `! rm /tmp/blah' to "reset the scratch space." To nex, we're always
  looking at a (headless) blank notebook. But reading a file in makes it
  complain, so we force-write and do a "w-! rm" dance.
* Verifying to myself that `q' really terminates execution, instead of
  going to the second quit command.