004 Early exit nex script if a command fails
We use a named-file convention of err.nex to see if an error occurred.
If the file is absent, then `read' fails and the rest of the nex script
is not run.

If we use this pattern, we must first check for and delete the err.nex
file at the top-level of the script.