007 Populate a template by reading it in & editing
Here we can replace `cat' with `r <file>', and it prints contents of the
file with nex instead of using a shell builtin.

Since we use :so[urce], I want to call these spellbooks and spells, etc.
But I also want to keep the .nex extension to denote they are
_transforms_ and not typical files. Then again, templates are pretty
much text files. (And I use ALL_CAPS for that, or trying to.)